MAINT: rename core score module
